Rename subpacakge `ingest` -> `data`
This is a legacy of valkyrie, which ingested supported data into a database.
