+++ tests/unit/test_itrf_converter.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import pandas as pd
 import pytest
 
-from iceflow.ingest.models import IceflowDataFrame
+from iceflow.data.models import IceflowDataFrame
 from iceflow.itrf.converter import _datetime_to_decimal_year, transform_itrf
